# openapi.yaml
openapi: 3.0.3
info:
  title: CaseFlow API
  description: |
    REST API for the CaseFlow application - a production-ready web app for 
    operations teams to upload CSV files, validate/clean data, and bulk-create cases.
  version: 1.0.0
  contact:
    email: hiring@skycladventures.com
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-deployment-url.com/api
    description: Production server
...
paths:
...
  /imports:
    get:
      summary: List imports
      tags: [Imports]
      security:
        - sessionAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of imports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportListResponse"
        "401":
          description: Unauthorized

    post:
      summary: Create import record
      tags: [Imports]
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportCreateInput"
      responses:
        "201":
          description: Import record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
        "401":
          description: Unauthorized
  /imports/{id}:
    get:
      summary: Get import details
      tags: [Imports]
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Import details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportDetailResponse"
        "401":
          description: Unauthorized
        "404":
          description: Import not found

    patch:
      summary: Update import status
      tags: [Imports]
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportUpdateInput"
      responses:
        "200":
          description: Import updated
        "401":
          description: Unauthorized
        "404":
          description: Import not found
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
...
    Category:
      type: string
      enum: [TAX, LICENSE, PERMIT]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    CaseStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, COMPLETED, REJECTED]
    ImportStatus:
      type: string
      enum: [PROCESSING, COMPLETED, FAILED]
...
    Case:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        applicantName:
          type: string
        dob:
          type: string
          format: date-time
        email:
          type: string
        phone:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        priority:
          $ref: "#/components/schemas/Priority"
        status:
          $ref: "#/components/schemas/CaseStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ImportCreateInput:
      type: object
      required: [filename, totalRows]
      properties:
        filename:
          type: string
        totalRows:
          type: integer
          minimum: 1
    ImportUpdateInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ImportStatus"
        successCount:
          type: integer
        failureCount:
          type: integer
        errorDetails:
          type: object
    ImportResponse:
      type: object
      properties:
        import:
          $ref: "#/components/schemas/Import"
    Import:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        totalRows:
          type: integer
        successCount:
          type: integer
        failureCount:
          type: integer
        status:
          $ref: "#/components/schemas/ImportStatus"
        createdAt:
          type: string
          format: date-time
    ImportListResponse:
      type: object
      properties:
        imports:
          type: array
          items:
            $ref: "#/components/schemas/Import"
        nextCursor:
          type: string
        hasMore:
          type: boolean
    ImportDetailResponse:
      type: object
      properties:
        import:
          allOf:
            - $ref: "#/components/schemas/Import"
            - type: object
              properties:
                errorDetails:
                  type: object
                cases:
                  type: array
                  items:
                    $ref: "#/components/schemas/Case"
...